reduce complexity of plugin manager
