+++ cognite/neat/plugins/_manager.py
@@ -1,17 +1,11 @@
 """Plugin manager for external plugins."""
 
 from importlib import metadata
-from typing import Any, ClassVar, TypeAlias, TypeVar, cast
+from typing import Any, ClassVar
 
+from ._base import NeatPlugin
+from ._data_model import DataModelImporterPlugin, DataModelTransformerPlugin
 from ._issues import PluginDuplicateError, PluginError, PluginLoadingError
-from .data_model.importers import DataModelImporterPlugin
-from .data_model.transformers._base import DataModelTransformerPlugin
-
-#: Type alias for all supported plugin types
-NeatPlugin: TypeAlias = DataModelImporterPlugin | DataModelTransformerPlugin
-
-# Generic type variable for plugin types
-T_NeatPlugin = TypeVar("T_NeatPlugin", bound=NeatPlugin)
 
 
 class Plugin:
@@ -50,7 +44,7 @@ class PluginManager:
     def __init__(self, plugins: dict[tuple[str, type[NeatPlugin]], Any]) -> None:
         self._plugins = plugins
 
-    def get(self, name: str, type_: type[T_NeatPlugin]) -> type[T_NeatPlugin]:
+    def get(self, name: str, type_: type[NeatPlugin]) -> type[NeatPlugin]:
         """
         Returns desired plugin
 
@@ -59,8 +53,8 @@ class PluginManager:
             type_ (type): The type of the plugin.
         """
         try:
-            plugin_class = self._plugins[(name, cast(type[NeatPlugin], type_))]
-            return cast(type[T_NeatPlugin], plugin_class)
+            plugin_class = self._plugins[(name, type_)]
+            return plugin_class
         except KeyError:
             raise PluginError(plugin_name=name, plugin_type=type_.__name__) from None
