--- cognite/neat/plugins/data_model/transformers/__init__.py
@@ -1,5 +1,0 @@
-from ._base import DataModelTransformerPlugin
-
-__all__ = [
-    "DataModelTransformerPlugin",
-]
